feat: AgentDefinition-driven sub-agent filtering and live bash/diff TUI rendering
- Add ToolManager::retain() for tool whitelist/blacklist filtering
- Add AgentDefinition to BackgroundSubAgentStart with model override
- build_filtered_tool_manager respects definition.tools/disallowed_tools
- resolve_kind_definition and resolve_spawn_agent_definition helpers
- plan_mode_required from definition propagates to AgentExecutionMode
- BackgroundSubAgentRecord gains model field for observability
- Split agent test module to agent_test.rs (1218 lines)
- TerminalCell gains output_deltas for live bash output streaming
- MessageCell renders inline diff via render_patch_preview for patches
- 31/31 tests pass
